requirements.txt
langchain
python-dotenv
langchain_mistralai
langchain_core
streamlit
langsmith
openai